requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
selenium>=4.25.0
anthropic>=0.39.0
pypdf>=4.0.0
reportlab>=4.0.0
python-docx>=1.1.0
holidays>=0.60
python-dateutil>=2.9.0
pydantic-settings>=2.5.0
aiofiles>=24.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
tzdata>=2024.1
beautifulsoup4>=4.12.0
